embedding_versioning/requirements.txt
numpy==1.26.4
faiss-cpu==1.7.4
pydantic==2.6.4
pytest==8.0.2
